// src/shared/types/enums.ts
export enum ErrorCode {
  EMAIL_ALREADY_EXISTS = "EMAIL_ALREADY_EXISTS",
  INVALID_DATA = "INVALID_DATA",
  USER_NOT_FOUND = "USER_NOT_FOUND",
  INVALID_OTP = "INVALID_OTP",
  INVALID_TOKEN = "INVALID_TOKEN",
  MISSING_TOKEN = "MISSING_TOKEN",
  UNVERIFIED_USER = "UNVERIFIED_USER",
  AUTH_FAILED = "AUTH_FAILED",
  EMAIL_SENDING_FAILED = "EMAIL_SENDING_FAILED",
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  INVALID_FILE_TYPE = "INVALID_FILE_TYPE",
  FILE_TOO_LARGE = "FILE_TOO_LARGE",
  FILE_UPLOAD_FAILED = "FILE_UPLOAD_FAILED",
  FAILED_TO_UPLOAD_IMAGE = "FAILED_TO_UPLOAD_IMAGE",
  USER_ALREADY_EXISTS = "USER_ALREADY_EXISTS",
  USER_ALREADY_VERIFIED = "USER_ALREADY_VERIFIED",
  RECORD_ALREADY_EXISTS = "RECORD_ALREADY_EXISTS",
  RECORD_NOT_FOUND = "RECORD_NOT_FOUND",
}